Feature/comments

+++ src/services/tickets/tickets.service.js
@@ -11,7 +11,10 @@ module.exports = {
     getById,
     create,
     update,
-    delete: _delete
+    delete: _delete,
+    createComment,
+    updateComment,
+    deleteComment
 }
 
 async function getAll(query) {
@@ -34,7 +37,7 @@ async function getById(ticket_id) {
     if (!ObjectId.isValid(ticket_id)) {
         throw 'Provided ticket_id is invalid'
     }
-    
+
     const ticket = await Ticket.findOne(new ObjectId(ticket_id))
 
     if (!ticket) {
@@ -44,13 +47,15 @@ async function getById(ticket_id) {
     return ticket
 }
 
-async function create(ticketParam) {
-    const ticket = new Ticket(ticketParam)
-    
+async function create(ticketParam, user) {
+    var ticket = new Ticket(ticketParam)
+    ticket.creator_id = user.uidNumber
+    ticket.creator_display_name = user.cn
+
     return await ticket.save()
 }
 
-async function update(ticket_id, ticketParam) {
+async function update(ticket_id, ticketParam, user_id) {
     if (!ObjectId.isValid(ticket_id)) {
         throw 'Provided ticket_id is invalid'
     }
@@ -59,15 +64,100 @@ async function update(ticket_id, ticketParam) {
 
     if (!existingTicket) throw 'Ticket not found'
 
+    if (existingTicket.creator_id != user_id) {
+        throw 'You do not have permissions to edit this comment'
+    }
+
     Object.assign(existingTicket, ticketParam)
 
     return await existingTicket.save()
 }
 
-async function _delete(ticket_id) {
+async function _delete(ticket_id, user_id) {
     if (!ObjectId.isValid(ticket_id)) {
         throw 'Ticket_id is invalid'
     }
 
+    const existingTicket = await Ticket.findOne(new ObjectId(ticket_id))
+
+    if (existingTicket.creator_id != user_id) {
+        throw 'You do not have permissions to delete this ticket'
+    }
+
     return await Ticket.deleteOne({ '_id': new ObjectId(ticket_id) })
+}
+
+async function createComment(ticket_id, commentParam, user) {
+    if (!ObjectId.isValid(ticket_id)) {
+        throw 'Provided ticket_id is invalid'
+    }
+
+    const ticket = await Ticket.findOne(new ObjectId(ticket_id))
+
+    if (!ticket) {
+        throw 'Ticket with given ID does not exist'
+    }
+
+    var comment = ticket.comments.create(commentParam)
+    comment.author_id = user.uidNumber
+    comment.author_display_name = user.cn
+
+    ticket.comments.push(comment)
+
+    return await ticket.save()
+}
+
+async function updateComment(comment_id, appParam, user_id) {
+    if (!ObjectId.isValid(comment_id)) {
+        throw 'Provided comment_id is invalid'
+    }
+
+    const ticket = await Ticket.findOne({ 'comments._id': new ObjectId(comment_id) })
+
+    if (!ticket) {
+        throw 'Comment with given ID does not exist'
+    }
+
+    var set = {}
+    for (var field in appParam) {
+        set['comments.$.' + field] = appParam[field]
+    }
+    set['comments.$.updated_at'] = new Date(Date.now())
+
+    const comment = ticket.comments.id(comment_id)
+
+    if (comment.author_id != user_id) {
+        throw 'You do not have permissions to edit this comment'
+    }
+
+    const updatedRecord = await Ticket.findOneAndUpdate({ 'comments._id': new ObjectId(comment_id) }, { $set: set }, { new: true })
+    if (!updatedRecord) {
+        throw 'Comment with given ID does not exist'
+    }
+
+    updatedRecord.save()
+
+    return updatedRecord
+}
+
+async function deleteComment(comment_id, user_id) {
+    if (!ObjectId.isValid(comment_id)) {
+        throw 'Provided comment_id is invalid'
+    }
+
+    const ticket = await Ticket.findOne({ 'comments._id': new ObjectId(comment_id) })
+
+    if (!ticket) {
+        throw 'Comment with given ID does not exist'
+    }
+
+    const comment = ticket.comments.id(comment_id)
+
+    if (comment.author_id != user_id) {
+        throw 'You do not have permissions to delete this comment'
+    }
+
+    await comment.remove()
+
+    return ticket.save()
 }
